[refactor] 불필요한 dto 삭제 후 디렉토리 삭제 #30
