--- good2travel/src/main/java/com/spot/good2travel/common/fcm/dto/FcmMessage.java
@@ -1,30 +1,0 @@
-package com.spot.good2travel.common.fcm.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@RequiredArgsConstructor
-@AllArgsConstructor
-@Builder
-public class FcmMessage {
-    private boolean validateOnly;
-    private Message message;
-
-    @Builder
-    @AllArgsConstructor
-    @Getter
-    public static class Message {
-        private Notification notification;
-        private String token;
-    }
-
-    @Builder
-    @AllArgsConstructor
-    @Getter
-    public static class Notification {
-        private String title;
-        private String body;
-    }
-
-}
